# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sicherheitsdienst-Tool API
  version: 1.0.0
  description: API v1 für Auth, Mitarbeiter, Sites, Schichten, Zuweisungen, Zeiterfassung, Vorfälle, Benachrichtigungen.
servers:
  - url: http://localhost:3000/api/v1
  - url: http://localhost:3001/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [success, code, message]
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: "Eingaben sind ungültig."
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Muss eine gültige E-Mail sein."
              code:
                type: string
                nullable: true
                example: "invalid_email"
      required: [message]
...
    DevicePlatform:
      type: string
      enum: [IOS, ANDROID, WEB]
    DeviceToken:
      type: object
      properties:
        platform:
          $ref: '#/components/schemas/DevicePlatform'
        token:
          type: string
        isActive:
          type: boolean
        notificationsEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [platform, token, isActive, notificationsEnabled, createdAt, updatedAt]
    DeviceTokenCreate:
      type: object
      properties:
        platform:
          $ref: '#/components/schemas/DevicePlatform'
        token:
          type: string
      required: [platform, token]
    DeviceTokenUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        notificationsEnabled:
          type: boolean
      additionalProperties: false
...
  responses:
    BadRequest:
      description: Ungültige Anfrage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: BAD_REQUEST
            message: Anfrageparameter sind ungültig.
    Unauthorized:
      description: Nicht autorisiert
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: UNAUTHORIZED
            message: Token fehlt oder ist ungültig.
...
    NotFound:
      description: Nicht gefunden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: NOT_FOUND
            message: Ressource nicht gefunden.
...
    InternalServerError:
      description: Serverfehler
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: INTERNAL_SERVER_ERROR
            message: Unerwarteter Fehler.
    ServiceUnavailable:
      description: Dienst nicht verfügbar
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            code: SERVICE_UNAVAILABLE
            message: Wartung oder temporärer Ausfall.
...
paths:
...
  /push/tokens:
    get:
      operationId: listMyTokens
      summary: List my push tokens
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/DeviceToken' }
                required: [success, data]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500':
          description: Push-Provider-Fehler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                success: false
                code: INTERNAL_SERVER_ERROR
                message: Push-Dienst vorübergehend nicht erreichbar.
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      operationId: registerToken
      summary: Register or update a device token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceTokenCreate' }
            examples:
              sample:
                value:
                  platform: ANDROID
                  token: fcm-xxx
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/DeviceToken' }
                required: [success, data]
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: Validierungsfehler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationError' }
              example:
                message: "Validierungsfehler."
                errors:
                  - field: platform
                    message: Muss einer der Werte IOS/ANDROID/WEB sein.
                  - field: token
                    message: Darf nicht leer sein.
        '500':
          description: Push-Provider-Fehler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                success: false
                code: INTERNAL_SERVER_ERROR
                message: Push-Dienst vorübergehend nicht erreichbar.
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /push/tokens/{token}:
    parameters:
      - in: path
        name: token
        required: true
        schema: { type: string }
    put:
      operationId: updateMyToken
      summary: Update a device token (status flags)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceTokenUpdate' }
            examples:
              disable:
                value:
                  isActive: false
                  notificationsEnabled: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/DeviceToken' }
                required: [success, data]
        '400':
          description: Ungültige Anfrage
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                code: BAD_REQUEST
                message: "Ungültige Flags übergeben."
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422':
          description: Validierungsfehler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationError' }
              example:
                message: "Validierungsfehler."
                errors:
                  - field: isActive
                    message: Muss boolean sein.
                  - field: notificationsEnabled
                    message: Muss boolean sein.
        '500':
          description: Push-Provider-Fehler
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                success: false
                code: INTERNAL_SERVER_ERROR
                message: Push-Dienst vorübergehend nicht erreichbar.
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    delete:
      operationId: deleteMyToken
      summary: Delete a device token
      security:
        - BearerAuth: []
      responses:
        '204': { description: No Content }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...